Implement part 1 of day 7 of AoC 2017; add lifetime to `ParseInput` trait
Since the `day07::Door` only holds references to the `input` string
(substrings which are found through regex matching) it has to carry a
generic lifetime and `ParseInput` has to be implemented such that its
lifetime matches that of the `input` string. For all the other `Door`
impls, the lifetime needed to be added as well to the `ParseInput`
impl, but there it doesn't matter, so `'_` can be used.