Add CLI option to skip computation of answers for solved puzzles
This is intended in case any solutions end up taking very long to
compute. Usually that's not the case for AoC problems when they are
solved efficiently (and usually they are not solvable without excessive
resources otherwise) but you never know. It might become annoying to
have to wait every time for those past solutions to be recalculated on
every single run. Hence, `--skip-solved` will only calculate answers
then the correct one has not yet been submitted to the server.