fix: No longer require parsing wrong guess from API responses
At some point along the way, AoC must have stopped including the wrong
guess in their responses. This caused wrong guesses to still be
classified as `IncorrectOther` but the information whether the guess
was too high or too low was lost.

This has been fixed by no longer attempting to parse out the wrong
guess. There really wasn't much of a benefit to this anyway as we know
that our guess was, so it really only served as a sanity check of sorts.