feat: Replace use of `async-trait` crate with AFIT
Since `JoinSet::spawn` requires the futures to be `Send`, we cannot use
plain vanilla AFIT, but need to express the `Send` bound.

This could have been achieved by using RPITIT with `impl Future + Send`,
but that would have baked in this restriction into the trait definition.
Instead, we pull in the `trait_variant` crate to generate a version with
`Send` bounds. Unlike `async_trait`, this comes at no runtime overhead
as it is no longer necessary to box the futures.