fix: remove unnecessary cast
For some reason, Clippy flagged this correctly in CI but not in my
local build. No idea why. I could not get Clippy to flag this, even
after ensuring that Rust/Clippy version, architecture, etc. match the
config on GH Actions.

Either way, the lint is correct and I've now fixed the issue.