feat: faster part 2 solution of day 22
Instead of iterating over all `[-9..=9;4]`, memorize the prices of the
first occurence of each 4-tuple of price changes in one go, then sum
those over all the starting values. This is far more efficient as it
allows iterating over the prices in a single pass, at the expense of
memory.