refactor: add higher-order function `cached` as sole public API of `cache` module
This is similar in spirit to the `@cache` decorator in the Python
`functools` module. You can easily create a function which memoizes its
argument and return value by wrapping it in `cached`. The function will
even accept a second argument `recurse` which effectively allows
recursively calling through the cache layer.