refactor: generalize day 12 solution a bit
Knowing that all problems can be conclusively judged irrespective of the
actual present shapes, it is no longer necessary to restrict the
solution to known shapes by asserting on a specific hash. All that's
needed is to extract the areas occupied by each present.

For different inputs, they might still be problems which cannot be
conclusively judged based on the implemented rules (though I suspect
that all inputs share this quality) but then my solution would still
produce a proper error, rather than a wrong answer.